refactor(structure): move all the src files into a src directory
