+++ src/components/organisms/StatsPanel/StatsPanel.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/variables.scss';
+@import 'styles/variables.scss';
 
 .title {
 	text-align: center;
