+++ src/components/organisms/StatsPanel/StatsPanel.tsx
