+++ src/components/templates/Footer/Footer.module.scss
