+++ src/models/RateSwitchboardState.ts
