+++ src/models/RedeemLogicForwardState.ts
