+++ src/pages/contract/summary/ClaimButton.tsx
@@ -4,13 +4,13 @@ import { useContext, useState } from 'react';
 import { AnchorProvider } from '@project-serum/anchor';
 import { useConnection, useWallet } from '@solana/wallet-adapter-react';
 import { PublicKey } from '@solana/web3.js';
-import { settle } from 'api/otc-state/settle';
+import { claim } from 'api/otc-state/claim';
 import ButtonPill from 'components/atoms/ButtonPill/ButtonPill';
+import { TxHandlerContext } from 'components/providers/TxHandlerProvider';
 import { PlusIcon } from 'evergreen-ui';
 import { useGetFetchOTCStateQuery } from 'hooks/useGetFetchOTCStateQuery';
-import { TxHandlerContext } from 'providers/TxHandlerProvider';
 
-export const SettleButton = ({ otcStatePubkey }: { otcStatePubkey: string }) => {
+export const ClaimButton = ({ otcStatePubkey, isBuyer }: { otcStatePubkey: string; isBuyer: boolean }) => {
 	const { connection } = useConnection();
 	const wallet = useWallet();
 	const txHandler = useContext(TxHandlerContext);
@@ -19,10 +19,10 @@ export const SettleButton = ({ otcStatePubkey }: { otcStatePubkey: string }) => 
 	const rateStateQuery = useGetFetchOTCStateQuery(provider, otcStatePubkey);
 	const [isLoading, setIsLoading] = useState(false);
 
-	const onSettleClick = async () => {
+	const onClaimClick = async () => {
 		try {
 			setIsLoading(true);
-			const tx = await settle(provider, new PublicKey(otcStatePubkey));
+			const tx = await claim(provider, new PublicKey(otcStatePubkey));
 			await txHandler.handleTxs(tx);
 		} catch (err) {
 			console.log(err);
@@ -32,9 +32,21 @@ export const SettleButton = ({ otcStatePubkey }: { otcStatePubkey: string }) => 
 		}
 	};
 
-	if (rateStateQuery?.data === undefined || !rateStateQuery?.data?.isSettlementAvailable()) {
+	if (isBuyer) {
+		if (rateStateQuery?.data === undefined || !rateStateQuery?.data?.isClaimSeniorAvailable(wallet.publicKey)) {
+			return <></>;
+		}
+	} else if (rateStateQuery?.data === undefined || !rateStateQuery?.data?.isClaimJuniorAvailable(wallet.publicKey)) {
 		return <></>;
 	}
 
-	return <ButtonPill mode="info" text="Settle" onClick={onSettleClick} icon={<PlusIcon />} loading={isLoading} />;
+	return (
+		<ButtonPill
+			mode={isBuyer ? 'success' : 'error'}
+			text={isBuyer ? 'Claim buyer' : 'Claim seller'}
+			onClick={onClaimClick}
+			icon={<PlusIcon />}
+			loading={isLoading}
+		/>
+	);
 };
