+++ src/pages/contract/summary/DepositButton.tsx
@@ -4,13 +4,13 @@ import { useContext, useState } from 'react';
 import { AnchorProvider } from '@project-serum/anchor';
 import { useConnection, useWallet } from '@solana/wallet-adapter-react';
 import { PublicKey } from '@solana/web3.js';
-import { claim } from 'api/otc-state/claim';
-import ButtonPill from 'components/atoms/ButtonPill/ButtonPill';
 import { PlusIcon } from 'evergreen-ui';
+import { deposit } from 'api/otc-state/deposit';
+import ButtonPill from 'components/atoms/ButtonPill/ButtonPill';
 import { useGetFetchOTCStateQuery } from 'hooks/useGetFetchOTCStateQuery';
-import { TxHandlerContext } from 'providers/TxHandlerProvider';
+import { TxHandlerContext } from 'components/providers/TxHandlerProvider';
 
-export const ClaimButton = ({ otcStatePubkey, isBuyer }: { otcStatePubkey: string; isBuyer: boolean }) => {
+export const DepositButton = ({ otcStatePubkey, isBuyer }: { otcStatePubkey: string; isBuyer: boolean }) => {
 	const { connection } = useConnection();
 	const wallet = useWallet();
 	const txHandler = useContext(TxHandlerContext);
@@ -19,10 +19,10 @@ export const ClaimButton = ({ otcStatePubkey, isBuyer }: { otcStatePubkey: strin
 	const rateStateQuery = useGetFetchOTCStateQuery(provider, otcStatePubkey);
 	const [isLoading, setIsLoading] = useState(false);
 
-	const onClaimClick = async () => {
+	const onDepositClick = async () => {
 		try {
 			setIsLoading(true);
-			const tx = await claim(provider, new PublicKey(otcStatePubkey));
+			const tx = await deposit(provider, new PublicKey(otcStatePubkey), isBuyer);
 			await txHandler.handleTxs(tx);
 		} catch (err) {
 			console.log(err);
@@ -33,18 +33,18 @@ export const ClaimButton = ({ otcStatePubkey, isBuyer }: { otcStatePubkey: strin
 	};
 
 	if (isBuyer) {
-		if (rateStateQuery?.data === undefined || !rateStateQuery?.data?.isClaimSeniorAvailable(wallet.publicKey)) {
+		if (rateStateQuery?.data === undefined || !rateStateQuery?.data?.isDepositBuyerAvailable(wallet.publicKey)) {
 			return <></>;
 		}
-	} else if (rateStateQuery?.data === undefined || !rateStateQuery?.data?.isClaimJuniorAvailable(wallet.publicKey)) {
+	} else if (rateStateQuery?.data === undefined || !rateStateQuery?.data?.isDepositSellerAvailable(wallet.publicKey)) {
 		return <></>;
 	}
 
 	return (
 		<ButtonPill
-			mode={isBuyer ? 'success' : 'error'}
-			text={isBuyer ? 'Claim buyer' : 'Claim seller'}
-			onClick={onClaimClick}
+			mode="success"
+			text={isBuyer ? 'Long' : 'Short'}
+			onClick={onDepositClick}
 			icon={<PlusIcon />}
 			loading={isLoading}
 		/>
