+++ src/pages/contract/summary/WithdrawButton.tsx
@@ -4,13 +4,13 @@ import { useContext, useState } from 'react';
 import { AnchorProvider } from '@project-serum/anchor';
 import { useConnection, useWallet } from '@solana/wallet-adapter-react';
 import { PublicKey } from '@solana/web3.js';
-import { deposit } from 'api/otc-state/deposit';
+import { withdraw } from 'api/otc-state/withdraw';
 import ButtonPill from 'components/atoms/ButtonPill/ButtonPill';
-import { PlusIcon } from 'evergreen-ui';
+import { TxHandlerContext } from 'components/providers/TxHandlerProvider';
+import { MinusIcon } from 'evergreen-ui';
 import { useGetFetchOTCStateQuery } from 'hooks/useGetFetchOTCStateQuery';
-import { TxHandlerContext } from 'providers/TxHandlerProvider';
 
-export const DepositButton = ({ otcStatePubkey, isBuyer }: { otcStatePubkey: string; isBuyer: boolean }) => {
+export const WithdrawButton = ({ otcStatePubkey, isBuyer }: { otcStatePubkey: string; isBuyer: boolean }) => {
 	const { connection } = useConnection();
 	const wallet = useWallet();
 	const txHandler = useContext(TxHandlerContext);
@@ -19,10 +19,10 @@ export const DepositButton = ({ otcStatePubkey, isBuyer }: { otcStatePubkey: str
 	const rateStateQuery = useGetFetchOTCStateQuery(provider, otcStatePubkey);
 	const [isLoading, setIsLoading] = useState(false);
 
-	const onDepositClick = async () => {
+	const onWithdrawClick = async () => {
 		try {
 			setIsLoading(true);
-			const tx = await deposit(provider, new PublicKey(otcStatePubkey), isBuyer);
+			const tx = await withdraw(provider, new PublicKey(otcStatePubkey));
 			await txHandler.handleTxs(tx);
 		} catch (err) {
 			console.log(err);
@@ -33,20 +33,12 @@ export const DepositButton = ({ otcStatePubkey, isBuyer }: { otcStatePubkey: str
 	};
 
 	if (isBuyer) {
-		if (rateStateQuery?.data === undefined || !rateStateQuery?.data?.isDepositBuyerAvailable(wallet.publicKey)) {
+		if (rateStateQuery?.data === undefined || !rateStateQuery?.data?.isWithdrawSeniorAvailable(wallet.publicKey)) {
 			return <></>;
 		}
-	} else if (rateStateQuery?.data === undefined || !rateStateQuery?.data?.isDepositSellerAvailable(wallet.publicKey)) {
+	} else if (rateStateQuery?.data === undefined || !rateStateQuery?.data?.isWithdrawJuniorAvailable(wallet.publicKey)) {
 		return <></>;
 	}
 
-	return (
-		<ButtonPill
-			mode="success"
-			text={isBuyer ? 'Long' : 'Short'}
-			onClick={onDepositClick}
-			icon={<PlusIcon />}
-			loading={isLoading}
-		/>
-	);
+	return <ButtonPill mode="error" text="Withdraw" onClick={onWithdrawClick} icon={<MinusIcon />} loading={isLoading} />;
 };
