+++ src/utils/stringHelpers.ts
